Add new `Rails/InverseOf` cop (#5045)
This cop looks for has_(one|many) and belongs_to associations where
ActiveRecord can't automatically determine the inverse association
because of a scope or the options used. This can result in unnecessary
queries in some circumstances. `:inverse_of` must be manually specified
for associations to work in both ways, or set to `false` to opt-out.

At Shopify we have tests that track SQL queries of critical paths, and fixing
these associations resulted in less database calls.