Add `httpdate` to accepted Rails time zone methods (#5761)
This format is required for certain header fields (e.g.
"If-Modified-Since"); which may be set in tests/specs. Using
[`httpdate`](https://ruby-doc.org/stdlib-2.5.0/libdoc/time/rdoc/Time.html#method-i-httpdate)
converts the time to a string using the HTTP-date format defined by [RFC
2616](https://www.w3.org/Protocols/rfc2616/rfc2616-sec3.html#sec3.3).
This method always converts the time to UTC.

The current Rubocop would require one of the following as a workaround:

  ```ruby
  Time.zone.now.httpdate
  Time.now.utc.httpdate
  Time.at(0).utc.httpdate
  ```

While this works, it is redundant and requires a unnecessary conversion
call. This change would allow:

  ```ruby
  Time.now.httpdate
  Time.at(0).httpdate
  ```