In Rails >= 5, use optional: true, not required: false (#6596)
The `required` option on belongs_to associations has been deprecated,
instead the recommended (and default) approach is to set this to true by
default and disable it using `optional: true`.

The developer can override this default globally, however, so we can't
say that things like `optional: false` are superfluous, or indeed that
`required: true` isn't necessary.

This change therefore only makes the known-to-be-valid correction from
`required: false` to `optional: true` and registers an offence for any
instances of `required`

See: https://guides.rubyonrails.org/5_0_release_notes.html