Allow model associations to reference class_name: self.name
When used in concern mixins, the class name that the association is included in
may be declared dynamically, instead of a static string.

This now only verifies that the `class_name:` option is not referencing a
constant, so other (dynamic) means of passing a string to the association
reflection can be supported.