Add new Rails/IndexBy and Rails/IndexWith cops
Active Support provides `Enumerable#index_by`, which creates a hash by
yielding each element and using the result as the key and the original
element as the value, and `Enumerable#index_with`, which uses the result
as the value and the original element as the key.

These cops identify and correct some patterns for doing the same thing
using `each_with_object`, `map { ... }.to_h`, or `Hash[map { ... }]`.

Co-authored-by: David Judd <djudd@stripe.com>