[Fix #215] Fix a false positive for `Rails/UniqueValidationWithoutIndex`
Fixes #215.

This PR fixes a false positive for `Rails/UniqueValidationWithoutIndex`
when using Expression Indexes for PostgreSQL.

Since it is difficult to predict a kind of string described in Expression Indexes,
this implementation judges based on whether or not an Expression Indexes include
a column name.
There is a possibility of false negative, but in most cases I expect it works.

## References

- https://www.postgresql.org/docs/12/indexes-expressional.html
- https://github.com/rails/rails/commit/edc2b7718725016e988089b5fb6d6fb9d6e16882