[Fix #238] Rails/IndexBy invalid syntax when .to_h is separated by a newline
I believe that the test cases included in this commit express well the
problem I'm trying to address, but running the auto-correction against
the following snippet of code:

```rb
x.map { |el| [el.to_sym, el] }.
  to_h
```

Would yield:

```rb
x.index_by { |el| el.to_sym }. # notice the ending "dot"
```

The output isn't parseable/valid Ruby. I belive the problem has to do
with the fact that correction code assumes that the `.to_h` statement is
always grouped together (which might not be always true).

The proposed solution was to try to compute the number of trailing characters
to be stripped by subtracting the end position of the map block (the `}`
character) from end position of the whole expression.