Handle undefined DidYouMean constants in UnknownEnv cop
The rubocop gem itself has a closed issue here:

https://github.com/rubocop-hq/rubocop/issues/7979

It details the fact that the DidYouMean constants are sometimes not
loaded even though the DidYouMean gem has been a part of Ruby since 2.3,
and should be required by default. The rubocop gem maintainers
determined this to be an upstream issue.

Here are some other relevant links to issues/PRs:

https://github.com/ruby/did_you_mean/issues/117
https://github.com/rubocop-hq/rubocop/pull/8143

My solution is mostly a band-aid fix to handle this problem until it is
fixed - use the DidYouMean spell checker if it is defined, otherwise
skip spell checking if it is not.