Handle to_h with block in IndexBy and IndexWith
Since Ruby 2.6, `Enumerable#to_h` accepts a block which is called with
each element and must return a key and value to include in the new hash.

If the element is always returned as either the key or the value, using
`Enumerable#index_by` or `Enumerable#index_with` is faster and clearer.

This is similar to the existing logic that detects `map { ... }.to_h`.