Fixes auto-correct of content_tag with hyphenated tag name
Allow auto-correction of tag names with hyphen, e.g. `trix-editor` to
`tag.trix_editor` instead of correcting to deprecated
`tag('trix-editor')` syntax.

The deprecated syntax does not support a content parameter, so the
arguments in the corrected code were incorrect when correcting a
method name with a hyphen and content, such as
`content_tag('trix-editor' 'my content!')`