Add cop to check ActiveModel errors manipulation as hash directly.
These are deprecated in Rails 6.1 and will be removed in Rails 7.
See https://github.com/rails/rails/pull/32313 for details.

The cop acts in two modes:

For files under `/models` directory, any `errors` call,
whether with receiver or not, will be checked.
For general files, only `errors` calls with receivers will be checked.

E.g. `errors[:bar] = []` is without receiver.
It will record an offense if it is a model file.
It will not record an offense if it is other general fie.

This is to reduce false-positives,
since other classes may also have a `errors` method.