[Fix #591] Add `change_column` check to `Rails/ReversibleMigration`
`change_column` is not on the list of the migration definitions supported by the `change` method. This is easy to miss as there are a couple `change_column_*` methods that are supported.

When trying to rollback a migration that uses `change_column` in the `change` method, the rollback will fail.  This change adds the check to `Rails/ReversibleMigration`.