[Fix #664] Fix a false positive for `Rails/MigrationClassName` cop
Fixes #664.

This PR fixes a false positive for `Rails/MigrationClassName` when
`ActiveSupport::Inflector` is applied to the class name and the case
is different.

Occasionally user want to treat a word like `Oauth` as` OAuth`.
This PR is case insensitive for comparisons between migration class name
and filename. It's may not a big deal because it's based on filename.