[Fix #841] Fix an error for `Rails/ActionOrder`
Fixes #841.

This PR fixes an error for `Rails/ActionOrder`
when using unconventional order of multiple actions.

With this change, autocorrection blank line positions are not as expected,
so they are left to autocorrection of `Layout/EmptyLineBetweenDefs` and
`Layout/EmptyLinesAroundClassBody` cops.

This PR aims to fix the error, that it may be a better autocorrection later.