Fix a false positive for `Rails/WhereMissing` when `left_joins(:foo)` and `where(foos: {id: nil})` separated by `or`, `and`
This PR is fix a false positive for `Rails/WhereMissing` when `left_joins(:foo)` and `where(foos: {id: nil})` separated by `or`, `and`.

It is necessary to consider the following cases, which are separated by `or` or `and`.

```ruby
Foo.left_joins(:foo).or(Foo.where(foos: {id: nil}))
Foo.where(foos: {id: nil}).and(Foo.left_joins(:foo))
```