[Fix #858] Mark the autocorrect for Rails/Delegate unsafe
Fixes https://github.com/rubocop/rubocop-rails/issues/858

The autocorrect for Rails/Delegate is unsafe when a refinement defines
the receiver. Since `delegate` defines the method dynamically
in a separate source file, the refinement is not applied in
the context of the method definition. Thus, calling the method
will result in a NameError because the receiver is not found.