[Fix #34] Allow `CreateTableWithTimestamps` when using `id: false` and not include `timestamps`
Fixes #34.

This PR allows `CreateTableWithTimestamps` when using `id: false` and not include `timestamps`.

There are design arguments for both allowing and disallowing migration file has `timestamps`
when `id: false` (join table) . For the following reasons, allow it when `id: false`:

For example, when running `bin/rails g migration create_articles`, the migration file containing
`timestamps` is created by default. Perhaps editing `id: false` and removing `timestamps` will be
intentional. On the other hand, if it is not intentionally deleted, it will remain.

So I think that it is possible to allow the behavior that `timestamps` is not written for `id: false`.
This emphasizes respecting user's editing intentions.