Fix database adapter detection for nested config
When a database yml is configured with nested named databases (for
example: `primary`, `read`, `replica`, etc) the adapter detection fails
by only looking for a top-level `adapter` key in the hash.

This changes the detection to try the top-level location first, and then
also attempt to look for an `adapter` key in the first nested hash under
`development` as well.