[Fix #1058] Relax `Include` path for `Rails/FindBy` and `Rails/FindEach`
Fixes #1058.

This PR relaxes `Include` path for `Rails/FindBy` and `Rails/FindEach`.

`Rails/FindBy` was introduced in https://github.com/rubocop/rubocop/commit/6188d75.
This was before it was extracted to RuboCop Rails, which is probably why it was
targeting `app/models`. So, RuboCop Rails targeting Rails would not need it.

`Rails/FindEach` is the same:
https://github.com/rubocop/rubocop/commit/f2e26812