Enable `InternalAffairs/UndefinedConfig`
This works fine right now.

I did a bit of spelunking for `NilOrEmpty` in `Rails/Present` and I believe it is a copy-paste error.
Added together with `Rails/Blank` in https://github.com/rubocop/rubocop/pull/4133/ which does
contain this config value. There are no tests or docs for this config value.