[Fix #1442] Fix an incorrect behavior when using `AllCops: TargetRailsVersion`
This PR fixes an incorrect behavior when using `AllCops: TargetRailsVersion`.

This suppresses the following warning when `TargetRailsVersion` is defined on the user side:

```console
$ bundle exec rubocop
(snip)

Warning: AllCops does not support TargetRailsVersion parameter.

Supported parameters are:

  - RubyInterpreters
  - Include
  - Exclude
  - DefaultFormatter
  - DisplayCopNames
  - DisplayStyleGuide
  - StyleGuideBaseURL
  - DocumentationBaseURL
  - DocumentationExtension
  - ExtraDetails
  - StyleGuideCopsOnly
  - EnabledByDefault
  - DisabledByDefault
  - NewCops
  - UseCache
  - MaxFilesInCache
  - CacheRootDirectory
  - AllowSymlinksInCacheRootDirectory
  - TargetRubyVersion
  - ParserEngine
  - SuggestExtensions
  - ActiveSupportExtensionsEnabled
  - StringLiteralsFrozenByDefault
```

This is a dirty hack relying on a private constant to prevent the warning.
It should be updated to a better design in the future.

Fixes #1442.