Automate the process of GitHub release creation
Follow-up to https://github.com/rubocop/rubocop/blob/master/.github/workflows/github_release.yml.

This action will be triggered when a new tag is pushed and will
auto-fill the release notes using the relevant file.

The `rubocop-rails` repository has essentially the same structure as the `rubocop` repository,
so it should work as is.