Avoid warnings about redefined methods in EnforceSuperclass module
Fix #1465

This change avoids the printing of warnings about some methods
(`included`, `on_class`, and `on_send`) of the
`RuboCop::Cop::EnforceSuperclass` module being redefined. These warnings
are printed because, in addition to the definition of that module here
in `rubocop-rails`, that module is first [also defined in `rubocop`][1].
(The module has been extracted to `rubocop-rails`, but is also left in
`rubocop`, for backwards compatibility.)

[1]: https://github.com/rubocop/rubocop/blob/v1.73.2/lib/rubocop/cop/mixin/enforce_superclass.rb

This change avoids the warnings by first undefining the
`RuboCop::Cop::EnforceSuperclass` module (the one from `rubocop`) before
then redefining that module here in `rubocop-rails`.