Add ConstantInRoutes Cop
In applications with large route files, its common to have a desire to
extract regexes, constraints, etc. that are used multiple times to
prevent duplication. If the value is never going to change, extracting
the value to a constant would be a very normal thing to do in typical
Ruby classes/modules. However, the `config/routes.rb` file is
reloadable, so a constant defined there will causing a constant
redefinition warning when/if the routes are reloaded (typically during
development or test, especially if using Spring).

This commit adds a ConstantInRoutes Cop to prevent constants from being
defined in `config/routes.rb` files.