Add `abort` to the `Rails/Exit` cop
`abort` is very similar to `exit` in Ruby
([docs](https://docs.ruby-lang.org/en/3.4/Kernel.html#method-i-abort)),
but it takes a string message or Exception. Rails applications should
avoid its usage for the same reasons as they avoid `exit` and `exit!`.
Since they are used in the same way, I thought it belonged in the same
cop, but am glad to move it to a new one if that's preferred.